+++ scaleph-support/scaleph-system/src/main/java/cn/sliew/scaleph/system/snowflake/worker/DisposableWorkerIdAssigner.java
@@ -19,11 +19,11 @@
 package cn.sliew.scaleph.system.snowflake.worker;
 
 import cn.sliew.milky.common.util.JacksonUtil;
+import cn.sliew.scaleph.common.util.NetUtils;
 import cn.sliew.scaleph.dao.DataSourceConstants;
 import cn.sliew.scaleph.dao.entity.master.snowflake.SnowflakeWorkerNode;
 import cn.sliew.scaleph.dao.mapper.master.snowflake.SnowflakeWorkerNodeMapper;
 import cn.sliew.scaleph.system.snowflake.utils.DockerUtils;
-import cn.sliew.scaleph.system.snowflake.utils.NetUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomUtils;
 import org.springframework.transaction.annotation.Transactional;
