+++ scaleph-ui-react/src/constants/enum.ts
@@ -14,3 +14,28 @@ export enum FlinkJobType {
   SEATUNNEL = '2',
   FLINK_CDC = '3'
 }
+
+export enum ResourceLifecycleState {
+  CREATED = 'CREATED',
+  SUSPENDED = 'SUSPENDED',
+  UPGRADING = 'UPGRADING',
+  DEPLOYED = 'DEPLOYED',
+  STABLE = 'STABLE',
+  ROLLING_BACK = 'ROLLING_BACK',
+  ROLLED_BACK = 'ROLLED_BACK',
+  FAILED = 'FAILED'
+}
+
+export enum FlinkJobState {
+  INITIALIZING = 'INITIALIZING',
+  CREATED = 'CREATED',
+  RUNNING = 'RUNNING',
+  FAILING = 'FAILING',
+  FAILED = 'FAILED',
+  CANCELLING = 'CANCELLING',
+  CANCELED = 'CANCELED',
+  FINISHED = 'FINISHED',
+  RESTARTING = 'RESTARTING',
+  SUSPENDED = 'SUSPENDED',
+  RECONCILING = 'RECONCILING',
+}
